Implement bitmap fileoutput.

+++ src/util/dimension.h
@@ -1,8 +1,10 @@
 #ifndef RAYTRACER_DIMENSION_H
 #define RAYTRACER_DIMENSION_H
 
+#include <stdint.h>
+
 typedef struct {
-    double width, height;
+    int32_t width, height;
 } Dimension;
 
 #endif //RAYTRACER_DIMENSION_H
